updated code

+++ bin/setup.sh
@@ -27,9 +27,15 @@ echo "*** Installing Selenium and Chrome for BDD"
 sudo apt-get update
 sudo DEBIAN_FRONTEND=noninteractive apt-get install -y sqlite3 ca-certificates chromium-driver python3-selenium
 
-echo "*** Installing Python depenencies..."
-source ~/venv/bin/activate && python3 -m pip install --upgrade pip wheel
-source ~/venv/bin/activate && pip install -r requirements.txt
+echo "*** Checking if requirements.txt exists..."
+if [ -f "requirements.txt" ]; then
+  echo "*** Installing Python dependencies..."
+  source ~/venv/bin/activate && python3 -m pip install --upgrade pip wheel
+  source ~/venv/bin/activate && pip install -r requirements.txt
+else
+  echo "Error: requirements.txt not found. Please ensure it exists before running the script."
+  exit 1
+fi
 
 echo "**************************************************"
 echo " CI/CD Final Project Environment Setup Complete"
